+++ examples/circles_model/xml/0.xml
@@ -1,0 +1,104 @@
+<states>
+    <itno>1</itno>
+    <environment/>
+    <xagent>
+        <name>circle1</name>
+        <x>3</x>
+        <y>3</y>
+    </xagent>
+    <xagent>
+        <name>circle1</name>
+        <x>3.0999999</x>
+        <y>3.1999998</y>
+    </xagent>
+    <xagent>
+        <name>circle1</name>
+        <x>3.2</x>
+        <y>3.4000001</y>
+    </xagent>
+    <xagent>
+        <name>circle1</name>
+        <x>3.3</x>
+        <y>3.5999999</y>
+    </xagent>
+    <xagent>
+        <name>circle1</name>
+        <x>3.4000001</x>
+        <y>3.8000002</y>
+    </xagent>
+    <xagent>
+        <name>circle1</name>
+        <x>3.5</x>
+        <y>4.545</y>
+    </xagent>
+    <xagent>
+        <name>circle1</name>
+        <x>3.5999999</x>
+        <y>4.1999998</y>
+    </xagent>
+    <xagent>
+        <name>circle1</name>
+        <x>3.7</x>
+        <y>4.4000001</y>
+    </xagent>
+    <xagent>
+        <name>circle1</name>
+        <x>3.8</x>
+        <y>4.5999999</y>
+    </xagent>
+    <xagent>
+        <name>circle1</name>
+        <x>3.9000001</x>
+        <y>4.8000002</y>
+    </xagent>
+    <xagent>
+        <name>circle2</name>
+        <x>2</x>
+        <y>2</y>
+    </xagent>
+    <xagent>
+        <name>circle2</name>
+        <x>2.2545</x>
+        <y>2.545</y>
+    </xagent>
+    <xagent>
+        <name>circle2</name>
+        <x>2.4000001</x>
+        <y>2.645</y>
+    </xagent>
+    <xagent>
+        <name>circle2</name>
+        <x>2.5999999</x>
+        <y>1.9999999</y>
+    </xagent>
+    <xagent>
+        <name>circle2</name>
+        <x>2.8</x>
+        <y>2.024</y>
+    </xagent>
+    <xagent>
+        <name>circle2</name>
+        <x>3.064</x>
+        <y>2.545</y>
+    </xagent>
+    <xagent>
+        <name>circle2</name>
+        <x>3.2</x>
+        <y>2.540</y>
+    </xagent>
+    <xagent>
+        <name>circle2</name>
+        <x>3.4000001</x>
+        <y>2.5654</y>
+    </xagent>
+    <xagent>
+        <name>circle2</name>
+        <x>3.5999999</x>
+        <y>1.9999999</y>
+    </xagent>
+    <xagent>
+        <name>circle2</name>
+        <x>3.8000002</x>
+        <y>2.078</y>
+    </xagent>
+</states>
