+++ src/flamegpu/model/ModelDescription.cu
@@ -9,8 +9,6 @@
  */
 
 #include <flamegpu/model/ModelDescription.h>
-#include <flamegpu/io/statereader.h>
-#include <flamegpu/io/statewriter.h>
 
 ModelDescription::ModelDescription(const std::string model_name) : agents(), messages(), name(model_name), population() {}
 
@@ -31,39 +29,6 @@ void ModelDescription::addMessage(const MessageDescription &message) {
 void ModelDescription::addPopulation(AgentPopulation &pop)
 {
  	population.insert(PopulationMap::value_type(pop.getAgentName(), pop));
-}
-
-/** 
-* Initialise the simulation. Allocated host and device memory. Reads the initial agent configuration from XML.
-* @param input	XML file path for agent initial configuration
-*/
-void ModelDescription::initialise(const ModelDescription &model, const char* input)
-{
-	std::unique_ptr<StateReader> read_ ;
-
-	StateReader &read__(read_->create(model, input));
-	read__.setFileName(input);
-	read__.setModelDesc(model);
-	read__.parse();
-
-	// todo : move factory class to outside (later)
- //We could use an if condition here to find out which derived class to create
-/* 
-	xmlReader read_ (model);
-
-	read_.setFileName(input);
-	read_.setModelDesc(model);
-	read_.parse();
-*/
-}
-	
-
-
-void ModelDescription::outputXML(const ModelDescription &model, char* output)
-{
-	//read initial states
-	StateWriter statewrite_;
-	statewrite_.writeStates(model, output);
 }
 
 const AgentDescription& ModelDescription::getAgentDescription(const std::string agent_name) const{
